DF-5739 fix pair of warnings and test on StaticWebView - works
Former-commit-id: 6ec133801167449d815004ecf4c5f31437af5341

+++ Sources/Internal/Render/RenderState.cpp
@@ -248,7 +248,7 @@ void RenderState::Flush(RenderState * hardwareState) const
             
             uint32 minIndex = Min(currentTextureData.minmaxTextureIndex & 0x000000FF, hardwareTextureData.minmaxTextureIndex & 0x000000FF);
             uint32 maxIndex = Max((currentTextureData.minmaxTextureIndex & 0x0000FF00), (hardwareTextureData.minmaxTextureIndex & 0x0000FF00)) >> 8;
-            for(size_t i = minIndex; i <= maxIndex; ++i)
+            for(uint32 i = minIndex; i <= maxIndex; ++i)
             {
                 if(currentTextureData.textures[i] != hardwareTextureData.textures[i])
                 {
